Refactor into modules. (#62)
* Refactor into modules.

* Update docs.

* Fix spacing.

* Add utils module.

* Update CHANGELOG.
