+++ docs/requirements.txt
@@ -3,3 +3,5 @@ sphinx_rtd_theme==0.4.3
 docutils==0.16     # >0.17 doesn't render bullets
 numpy>=1.22      # so that default dtype are correctly rendered
 torch>=1.10
+torchvision>=0.15.2
+torchmetrics>=0.11.4
