+++ lensless/__init__.py
@@ -13,25 +13,25 @@ Package for designing, simulating, and imaging with lensless cameras.
 """
 
 
-from .recon import ReconstructionAlgorithm
-from .admm import ADMM
-from .gradient_descent import (
+from .recon.recon import ReconstructionAlgorithm
+from .recon.admm import ADMM
+from .recon.gd import (
     GradientDescent,
     NesterovGradientDescent,
     FISTA,
     GradientDescentUpdate,
 )
-from .sensor import VirtualSensor, SensorOptions
+from .hardware.sensor import VirtualSensor, SensorOptions
 
 try:
-    from .trainable_recon import TrainableReconstructionAlgorithm
-    from .unrolled_admm import UnrolledADMM
-    from .unrolled_fista import UnrolledFISTA
+    from .recon.trainable_recon import TrainableReconstructionAlgorithm
+    from .recon.unrolled_admm import UnrolledADMM
+    from .recon.unrolled_fista import UnrolledFISTA
 except Exception:
     pass
 
 try:
-    from .apgd import APGD, APGDPriors
+    from .recon.apgd import APGD, APGDPriors
 
     pycsou_available = True
 except Exception:
