+++ lensless/recon/apgd.py
@@ -7,11 +7,11 @@
 # #############################################################################
 
 
-from lensless.recon import ReconstructionAlgorithm
+from lensless.recon.recon import ReconstructionAlgorithm
 import inspect
 import numpy as np
 from typing import Optional
-from lensless.rfft_convolve import RealFFTConvolve2D as Convolver
+from lensless.recon.rfft_convolve import RealFFTConvolve2D as Convolver
 
 import pycsou.abc as pyca
 import pycsou.operator.func as func
