+++ profile/gradient_descent.py
@@ -1,7 +1,7 @@
 from diffusercam_gd import grad_descent, GradientDescentUpdate
-from lensless.gradient_descent import FISTA
-from lensless.apgd import APGD, APGDPriors
-from lensless.io import load_data
+from lensless.recon.gd import FISTA
+from lensless.recon.apgd import APGD, APGDPriors
+from lensless.utils.io import load_data
 import time
 import pathlib as plib
