+++ scripts/analyze_image.py
@@ -41,9 +41,9 @@ import click
 import cv2
 import numpy as np
 import matplotlib.pyplot as plt
-from lensless.util import rgb2gray
-from lensless.plot import plot_image, pixel_histogram, plot_cross_section, plot_autocorr2d
-from lensless.io import load_image
+from lensless.utils.image import rgb2gray
+from lensless.utils.plot import plot_image, pixel_histogram, plot_cross_section, plot_autocorr2d
+from lensless.utils.io import load_image
 
 
 @click.command()
