+++ scripts/on_device_capture.py
@@ -18,8 +18,9 @@ import cv2
 import numpy as np
 from time import sleep
 from PIL import Image
-from lensless.util import bayer2rgb, get_distro, rgb2gray, resize
-from lensless.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
+from lensless.hardware.utils import get_distro
+from lensless.utils.image import bayer2rgb, rgb2gray, resize
+from lensless.hardware.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
 from fractions import Fraction
 import time
