+++ scripts/sim/torch_dataset.py
@@ -2,8 +2,8 @@ from waveprop.dataset_util import SimulatedPytorchDataset
 import hydra
 from hydra.utils import to_absolute_path
 import os
-from lensless.io import load_psf
-from lensless.util import rgb2gray
+from lensless.utils.io import load_psf
+from lensless.utils.image import rgb2gray
 import torch
 import time
 import numpy as np
